requirements.txt
pandas 
matplotlib 
requests 
pytz 
google-generativeai 
streamlit
google-generativeai 
alpha_vantage       
Pillow
